+++ src/main/java/com/example/learnSpringBoot/exception/AnotherCustomException.java
@@ -1,0 +1,11 @@
+package com.example.learnSpringBoot.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class AnotherCustomException extends RuntimeException{
+
+	String message;
+}