+++ src/main/java/com/example/learnSpringBoot/exception/SomeCustomException.java
@@ -1,0 +1,20 @@
+package com.example.learnSpringBoot.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class SomeCustomException extends RuntimeException{
+	
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = 1L;
+	
+	private String userName;
+	private String message;
+	
+	
+
+}