+++ changeLog.md
@@ -1,4 +1,7 @@
-#Commit 3
+#Commit 3 [15-02-21]
 - Created controller
 - Used lombok for bean creation
 - Added live reload dependency
+
+#Commit 4 [15-03-21]
+- Exception Handling using @ControllerAdvice and @ExceptionHandler